About animal rescuer appplication.
Animal rescuer is my first application which was created like a homework at fasttrackit course. It's an application like tamagotchi type. From the start at this application user have to choose type of animal what he want to care about and start to feed, to give some medicine or just to play with his animal.